[AIRFLOW-4487] Move k8s executor from contrib folder to main project (#5261)
* Move k8s executor from contrib folder

Considering that the k8s executor is now fully supported by core
committers, we should move it from contrib to the primary executor
directory.
