+++ airflow/kubernetes/pod.py
@@ -69,6 +69,7 @@ class Pod:
     :param configmaps: A list containing names of configmaps object
         mounting env variables to the pod
     :type configmaps: list[str]
+
     """
     def __init__(
             self,
